# Add map cache in SliceStringLiteral