--- a/ast.go
+++ b/ast.go
@@ -100,6 +100,7 @@ func (n *NumberLiteral) Args() []string {
 
 type SliceStringLiteral struct {
 	Val []string
+	m   map[string]struct{}
 }
 
 // String returns a string representation of the literal.