--- a/evaluator.go
+++ b/evaluator.go
@@ -261,24 +261,18 @@ func applyIN(l, r Expr) (*BooleanLiteral, error) {
 	switch t := l.(type) {
 	case *StringLiteral:
 		var a string
-		var b []string
+		var b map[string]struct{}
+
 		a, err = getString(l)
 		if err != nil {
 			return nil, err
 		}
 
-		b, err = getSliceString(r)
-
+		b, err = getMapString(r)
 		if err != nil {
 			return nil, err
 		}
-
-		found = false
-		for _, e := range b {
-			if a == e {
-				found = true
-			}
-		}
+		_, found = b[a]
 	case *NumberLiteral:
 		var a float64
 		var b []float64
@@ -532,13 +526,20 @@ func getSliceNumber(e Expr) ([]float64, error) {
 	}
 }
 
-// getSliceString performs type assertion and returns []string value or error
-func getSliceString(e Expr) ([]string, error) {
+// getMapString performs type assertion and returns map[string]struct{} value or error
+func getMapString(e Expr) (map[string]struct{}, error) {
 	switch n := e.(type) {
 	case *SliceStringLiteral:
-		return n.Val, nil
+		if n.m != nil {
+			return n.m, nil
+		}
+		n.m = make(map[string]struct{})
+		for _, item := range n.Val {
+			n.m[item] = struct{}{}
+		}
+		return n.m, nil
 	default:
-		return []string{}, fmt.Errorf("Literal is not a slice of string: %v", n)
+		return nil, fmt.Errorf("Literal is not a slice of string: %v", n)
 	}
 }