Old Result
```
pkg: github.com/zhouzhuojie/conditions
BenchmarkLongSliceString-4        200000              8696 ns/op
PASS
ok      github.com/zhouzhuojie/conditions       3.050s

```

New Result
```
BenchmarkLongSliceString-4      10000000               104 ns/op
PASS
ok      github.com/zhouzhuojie/conditions       1.767s

```